add order created scenario and update readme file (#1)
* add order created case and update readme file

* add refactor including sendgrid service

* test circleci
